Simplifying first program

+++ 01_metody/src/main/java/pl/samouczekprogramisty/kursjava/methods/TemperatureCheck.java
@@ -2,6 +2,7 @@ package pl.samouczekprogramisty.kursjava.methods;
 
 public class TemperatureCheck {
     public static void main(String... args) {
+        // this is not the best type to store temperature, stay tuned for better one ;)
         int positiveTemperature = 10;
         int negativeTemperature = -10;
 
@@ -9,7 +10,7 @@ public class TemperatureCheck {
         isTemperaturePositive(negativeTemperature);
     }
 
-    private static boolean isTemperaturePositive(double temperature) {
+    private static boolean isTemperaturePositive(int temperature) {
         boolean isPositive = temperature > 0;
 
         if (isPositive) {
